+++ Common/Internal/Roslyn/RoslynInternals.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 #if !NETCOREAPP
 using System.Buffers;
 #endif
@@ -51,6 +50,8 @@ namespace MirrorSharp.Internal.Roslyn {
             if (roslynVersion.Major == 42 && roslynVersion.Minor == 42) {
                 // Try previous versions, in case CI is not on newest yet
                 assembly = GetAssemblyOrNullIfTypesFailToLoad(assembly)
+                        ?? GetAssemblyOrNullIfTypesFailToLoad(LoadInternalsAssemblySlow(new Version(4, 6)))
+                        ?? GetAssemblyOrNullIfTypesFailToLoad(LoadInternalsAssemblySlow(new Version(4, 5)))
                         ?? GetAssemblyOrNullIfTypesFailToLoad(LoadInternalsAssemblySlow(new Version(4, 4)))
                         ?? GetAssemblyOrNullIfTypesFailToLoad(LoadInternalsAssemblySlow(new Version(4, 3)))
                         ?? LoadInternalsAssemblySlow(new Version(4, 2));
@@ -73,7 +74,9 @@ namespace MirrorSharp.Internal.Roslyn {
 
         private static Assembly LoadInternalsAssemblySlow(Version roslynVersion) {
             var assemblyName = roslynVersion switch {
-                { Major: > 4 } or { Major: 4, Minor: >= 5 } => "MirrorSharp.Internal.Roslyn45.dll",
+                { Major: > 4 } or { Major: 4, Minor: >= 7 } => "MirrorSharp.Internal.Roslyn47.dll",
+                { Major: 4, Minor: >= 6 } => "MirrorSharp.Internal.Roslyn46.dll",
+                { Major: 4, Minor: >= 5 } => "MirrorSharp.Internal.Roslyn45.dll",
                 { Major: 4, Minor: >= 4 } => "MirrorSharp.Internal.Roslyn44.dll",
                 { Major: 4, Minor: >= 3 } => "MirrorSharp.Internal.Roslyn43.dll",
                 { Major: 4, Minor: 2 } => "MirrorSharp.Internal.Roslyn42.dll",