+++ Common/Properties/AssemblyInfo.cs
@@ -1,4 +1,3 @@
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // Setting ComVisible to false makes the types in this assembly not visible
@@ -8,26 +7,3 @@ using System.Runtime.InteropServices;
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("c130c962-d17f-4741-b2b4-d74263bf380a")]
-
-[assembly: InternalsVisibleTo("MirrorSharp.AspNetCore")]
-[assembly: InternalsVisibleTo("MirrorSharp.Benchmarks")]
-[assembly: InternalsVisibleTo("MirrorSharp.CSharp.Scripting")]
-[assembly: InternalsVisibleTo("MirrorSharp.Internal.Roslyn33")]
-[assembly: InternalsVisibleTo("MirrorSharp.Internal.Roslyn36")]
-[assembly: InternalsVisibleTo("MirrorSharp.Internal.Roslyn41")]
-[assembly: InternalsVisibleTo("MirrorSharp.Internal.Roslyn42")]
-[assembly: InternalsVisibleTo("MirrorSharp.Internal.Roslyn43")]
-[assembly: InternalsVisibleTo("MirrorSharp.Internal.Roslyn44")]
-[assembly: InternalsVisibleTo("MirrorSharp.Internal.Roslyn45")]
-[assembly: InternalsVisibleTo("MirrorSharp.FSharp")]
-[assembly: InternalsVisibleTo("MirrorSharp.VisualBasic")]
-[assembly: InternalsVisibleTo("MirrorSharp.Php")]
-[assembly: InternalsVisibleTo("MirrorSharp.IL")]
-[assembly: InternalsVisibleTo("MirrorSharp.Owin")]
-[assembly: InternalsVisibleTo("MirrorSharp.Testing")]
-[assembly: InternalsVisibleTo("MirrorSharp.Tests")]
-[assembly: InternalsVisibleTo("MirrorSharp.Tests.Roslyn44")]
-[assembly: InternalsVisibleTo("MirrorSharp.Tests.RoslynLatest")]
-[assembly: InternalsVisibleTo("SharpLab.Server")]
-[assembly: InternalsVisibleTo("SharpLab.Tests")]
-[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]