+++ WebAssets/src/testing/jest/setup-jest.ts
@@ -1,0 +1,6 @@
+import { toMatchImageSnapshot } from 'jest-image-snapshot';
+
+expect.extend({ toMatchImageSnapshot });
+
+jest.useFakeTimers();
+jest.setTimeout(3 * 60 * 1000);