+++ WebAssets/tests/types/keysim.d.ts
@@ -1,8 +1,0 @@
-declare module 'keysim' {
-    export class Keyboard {
-        static readonly US_ENGLISH: Keyboard;
-
-        dispatchEventsForInput(input: string, target: HTMLElement): void;
-        dispatchEventsForAction(input: string, target: HTMLElement): void;
-    }
-}